Apps fixed for VS 2015

+++ utils/common/common_utils.h
@@ -22,7 +22,7 @@
 
 
 #ifdef _WIN32
-typedef unsigned long long SOCKET;
+typedef unsigned int SOCKET;
 #else
 typedef int SOCKET;
 #endif
